Refactor functions in files
